pyproject.toml
[project]
name = "duties-of-the-hearts"
version = "0.1.0"
description = ""
authors = [
    {name = "Itay Zandbank",email = "itay@zmbq.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "python-docx>=1.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false